Integration card rename and deprecate existing one (#74)
* Deprecate existing integration card

* Rename new card to Integration card and update references

---------
