+++ packages/ui/domain-components/index.ts
@@ -1,5 +1,5 @@
 // codegen:start {preset: barrel, include: "./{*.{ts,tsx},*/index.{ts,tsx}}", exclude: "./**/*.{d,spec,test,fixture,gen,node}.{ts,tsx}"}
 export * from './ConnectorCard'
-export * from './IntegrationCard'
+export * from './DeprecatedIntegrationCard'
 export * from './ResourceCard'
 // codegen:end
