+++ packages/ui/stories/IntegrationCard.stories.tsx
@@ -1,11 +1,11 @@
 import {fn} from '@storybook/test'
 import logoGreenhouse from '../../../apps/web/public/_assets/logo-greenhouse.svg'
-import {ConnectionCard} from '../domain-components/ConnectionCard'
+import {IntegrationCard} from '../domain-components/IntegrationCard'
 
 // More on how to set up stories at: https://storybook.js.org/docs/writing-stories#default-export
 export default {
-  title: 'Example/ConnectionCard',
-  component: ConnectionCard,
+  title: 'Example/IntegrationCard',
+  component: IntegrationCard,
   parameters: {
     // Optional parameter to center the component in the Canvas. More info: https://storybook.js.org/docs/configure/story-layout
     layout: 'centered',
